+++ uk/co/petertribble/sphaero2/ClassicCutter.java
@@ -64,7 +64,7 @@ public class ClassicCutter extends JigsawCutter {
 	 */
 
 	int rows = (int) Math.round(Math.sqrt(prefPieces * height / width));
-	int columns = (int) Math.round(prefPieces / rows);
+	int columns = Math.round(prefPieces / rows);
 
 	jp.setMaximum(rows*columns);
 	jp.setValue(iprogress);
